Add files via upload
LRS.py - the script I wrote to obtain the LRS data and create a csv file called "nyc_lrs.csv" (not included here). 

"nyc_lrs_modified.csv" - I opened the original csv file in Tableau and modified it to include a unique identifier for each census tract, then I exported it to this file. (Not necessary to export it, but I was having a little issue with getting it to work and this seemed to help).

nyc_lrs_modified.csv+(Multiple Connections).csv - is the result of joining the modified file to the census tract shape file (also included here). 

I used this the above data to create the maps.